requirements.txt
opencv-python-headless<4.3
streamlit==1.16.0
mediapipe==0.8.9.1
numpy==1.19.3
Pillow==8.2.0
click==8.0.4
pyttsx3==2.90
scikit-learn==1.1.3